+++ app/client/library-modules/use-cases/tenant-application/CreateTenantApplicationUseCase.ts
@@ -1,51 +1,11 @@
 // app/client/library-modules/use-cases/tenant-application/TenantApplicationUseCase.ts
-import { insertTenantApplication, updateTenantApplicationStatus } from "/app/client/library-modules/domain-models/tenant-application/repositories/tenant-application-repository";
+import { insertTenantApplication } from "/app/client/library-modules/domain-models/tenant-application/repositories/tenant-application-repository";
 import { getPropertyById } from "/app/client/library-modules/domain-models/property/repositories/property-repository";
 import { Role } from "/app/shared/user-role-identifier";
 import { TenantApplicationStatus } from "/app/shared/api-models/tenant-application/TenantApplicationStatus";
 import { CreateTenantApplicationRequest, CreateTenantApplicationResponse } from "./models/TenantApplicationsData";
-  export function validateCreateTenantApplicationRequest(request: CreateTenantApplicationRequest): void {
-    if (!request.currentUser) {
-      throw new Error('User must be logged in to apply');
-    }
 
-    if (!request.propertyId) {
-      throw new Error('Property ID is required to create application');
-    }
-
-    if (!request.propertyLandlordId) {
-      throw new Error('Property landlord ID is required');
-    }
-  }
-
-  export function validateUserCanApply(authUser: any, currentUser: any, bookedInspections: Set<number>): void {
-    if (!currentUser) {
-      throw new Error('Please log in to apply for this property.');
-    }
-
-    if (authUser && authUser.role !== Role.TENANT) {
-      throw new Error('Only tenants can apply for properties.');
-    }
-
-    if (bookedInspections.size === 0) {
-      throw new Error('Please book an inspection before applying for this property.');
-    }
-  }
-
-  export function determineApplicantName(currentUser: any, profileData?: any): string {
-    if (profileData?.firstName && profileData?.lastName) {
-      return `${profileData.firstName} ${profileData.lastName}`;
-    }
-
-    if (currentUser?.tenantId) {
-      return `Tenant ${currentUser.tenantId.slice(-4)}`;
-    }
-
-    throw new Error('Only tenants can apply for properties. Invalid user type detected.');
-  }
-
-
-  export async function createTenantApplicationUseCase(request: CreateTenantApplicationRequest): Promise<CreateTenantApplicationResponse> {
+export async function createTenantApplicationUseCase(request: CreateTenantApplicationRequest): Promise<CreateTenantApplicationResponse> {
     try {
       // Validate user can apply
       validateUserCanApply(request.authUser, request.currentUser, request.bookedInspections);
@@ -84,3 +44,44 @@ import { CreateTenantApplicationRequest, CreateTenantApplicationResponse } from 
       };
     }
   }
+
+export function validateCreateTenantApplicationRequest(request: CreateTenantApplicationRequest): void {
+  if (!request.currentUser) {
+    throw new Error('User must be logged in to apply');
+  }
+
+  if (!request.propertyId) {
+    throw new Error('Property ID is required to create application');
+  }
+
+  if (!request.propertyLandlordId) {
+    throw new Error('Property landlord ID is required');
+  }
+}
+
+export function validateUserCanApply(authUser: any, currentUser: any, bookedInspections: Set<number>): void {
+  if (!currentUser) {
+    throw new Error('Please log in to apply for this property.');
+  }
+
+  if (authUser && authUser.role !== Role.TENANT) {
+    throw new Error('Only tenants can apply for properties.');
+  }
+
+  if (bookedInspections.size === 0) {
+    throw new Error('Please book an inspection before applying for this property.');
+  }
+}
+
+export function determineApplicantName(currentUser: any, profileData?: any): string {
+  if (profileData?.firstName && profileData?.lastName) {
+    return `${profileData.firstName} ${profileData.lastName}`;
+  }
+
+  if (currentUser?.tenantId) {
+    return `Tenant ${currentUser.tenantId.slice(-4)}`;
+  }
+
+  throw new Error('Only tenants can apply for properties. Invalid user type detected.');
+}
+
