+++ app/client/library-modules/use-cases/tenant-application/ProcessTenantApplicationUseCase.ts
@@ -6,22 +6,34 @@ import { Role } from "/app/shared/user-role-identifier";
 import { TenantApplicationStatus } from "/app/shared/api-models/tenant-application/TenantApplicationStatus";
 import { TenantApplication } from "/app/client/ui-modules/tenant-selection/types/TenantApplication";
 
-
-export async function acceptTenantApplication(applicationId: string): Promise<{ applicationId: string; status: TenantApplicationStatus }> {
+export async function acceptTenantApplicationUseCase(applicationId: string): Promise<{ applicationId: string; status: TenantApplicationStatus }> {
   await updateTenantApplicationStatus([applicationId], TenantApplicationStatus.ACCEPTED, 1);
   return { applicationId, status: TenantApplicationStatus.ACCEPTED };
 }
 
-export async function rejectTenantApplication(applicationId: string): Promise<{ applicationId: string; status: TenantApplicationStatus }> {
+export async function rejectTenantApplicationUseCase(applicationId: string): Promise<{ applicationId: string; status: TenantApplicationStatus }> {
   await updateTenantApplicationStatus([applicationId], TenantApplicationStatus.REJECTED, 1);
   return { applicationId, status: TenantApplicationStatus.REJECTED };
 }
-export async function sendAcceptedApplicationsToLandlord(
+
+export async function sendAcceptedApplicationsToLandlordUseCase(
   propertyId: string,
   propertyLandlordId: string,
-  propertyAddress: string,
+  streetNumber: string,
+  street: string,
+  suburb: string,
+  province: string,
+  postcode: string,
   acceptedApplications: TenantApplication[]
 ): Promise<{ success: boolean; propertyId: string; acceptedApplications: string[]; applicationCount: number; taskId: string }> {
+
+  if (!propertyId) {
+    throw new Error('Property ID is required to send applications to landlord');
+  }
+
+  if (!propertyLandlordId) {
+    throw new Error('Property landlord ID is required to send tenant to landlord');
+  }
 
   if (acceptedApplications.length === 0) {
     throw new Error('No accepted applications to send to landlord');
@@ -29,22 +41,32 @@ export async function sendAcceptedApplicationsToLandlord(
 
   // Create task for landlord
   const taskName = `Review ${acceptedApplications.length} Tenant Application(s)`;
-  const taskDescription = `Review ${acceptedApplications.length} accepted tenant application(s) for property at ${propertyAddress}. Applicants: ${acceptedApplications.map(app => app.name).join(', ')}`;
+  const taskDescription = `Review ${acceptedApplications.length} accepted tenant application(s) for property at ${streetNumber} ${street}, ${suburb}, ${province} ${postcode}. Applicants: ${acceptedApplications.map((app: TenantApplication) => app.name).join(', ')}`;
+  const dueDate = new Date();
+  dueDate.setDate(dueDate.getDate() + 7);
 
+  // Create task for landlord
   const taskResult = await apiCreateTaskForLandlord({
     name: taskName,
     description: taskDescription,
-    dueDate: calculateDueDate(7),
+    dueDate: dueDate,
     priority: TaskPriority.MEDIUM,
-    propertyAddress,
-    propertyId,
-    userType: Role.LANDLORD,
     userId: propertyLandlordId,
+    propertyAddress: `${streetNumber} ${street}, ${suburb}, ${province} ${postcode}`,
+    propertyId: propertyId,
+    userType: Role.LANDLORD,
   });
 
-  // Update applications to LANDLORD_REVIEW status
-  const acceptedApplicationIds = acceptedApplications.map(app => app.id);
-  await updateTenantApplicationStatus(acceptedApplicationIds, TenantApplicationStatus.LANDLORD_REVIEW, 2, taskResult);
+  // Update all accepted applications to LANDLORD_REVIEW status
+  const acceptedApplicationIds = acceptedApplications.map((app: TenantApplication) => app.id);
+  await updateTenantApplicationStatus(
+    acceptedApplicationIds,
+    TenantApplicationStatus.LANDLORD_REVIEW,
+    2,
+    taskResult // Passes the task ID to link applications to the task
+  );
+
+  console.log(`Successfully sent ${acceptedApplications.length} application(s) to landlord ${propertyLandlordId} for property ${propertyId}`);
 
   return {
     success: true,
