+++ app/client/ui-modules/tenant-selection/state/reducers/tenant-selection-slice.ts
@@ -4,17 +4,12 @@ import { TenantApplication } from "../../types/TenantApplication";
 import { TenantApplicationStatus } from "../../../../../shared/api-models/tenant-application/TenantApplicationStatus";
 import { FilterType } from "../../enums/FilterType";
 import { TenantSelectionUiState } from "../TenantSelectionUiState";
-import { apiCreateTaskForLandlord } from "/app/client/library-modules/apis/task/task-api";
-import { TaskPriority } from "/app/shared/task-priority-identifier";
-import {
-  insertTenantApplication,
-  updateTenantApplicationStatus
-} from "/app/client/library-modules/domain-models/tenant-application/repositories/tenant-application-repository";
 import { LoadTenantApplicationsUseCase } from "/app/client/library-modules/use-cases/tenant-application/LoadTenantApplicationsUseCase";
-import { getPropertyById } from "/app/client/library-modules/domain-models/property/repositories/property-repository";
 import {
-  acceptTenantApplication,
-  rejectTenantApplication } from "/app/client/library-modules/use-cases/tenant-application/ProcessTenantApplicationWorkflowUseCase";
+  acceptTenantApplicationUseCase,
+  rejectTenantApplicationUseCase,
+  sendAcceptedApplicationsToLandlordUseCase
+} from "../../../../library-modules/use-cases/tenant-application/ProcessTenantApplicationUseCase";
 import { createTenantApplicationUseCase } from "/app/client/library-modules/use-cases/tenant-application/CreateTenantApplicationUseCase";
 import { Role } from "/app/shared/user-role-identifier";
 const initialState: TenantSelectionUiState = {
@@ -76,14 +71,14 @@ export const createTenantApplicationAsync = createAsyncThunk(
 export const rejectTenantApplicationAsync = createAsyncThunk(
   "tenantSelection/rejectTenantApplication",
   async (applicationId: string) => {
-    return await rejectTenantApplication(applicationId);
+    return await rejectTenantApplicationUseCase(applicationId);
   }
 );
 
 export const acceptTenantApplicationAsync = createAsyncThunk(
   "tenantSelection/acceptTenantApplication",
   async (applicationId: string) => {
-    return await acceptTenantApplication(applicationId);
+    return await acceptTenantApplicationUseCase(applicationId);
   }
 );
 
@@ -95,14 +90,6 @@ export const sendAcceptedApplicationsToLandlordAsync = createAsyncThunk(
     const { propertyId, propertyLandlordId, streetNumber, street, suburb, province, postcode } = state.propertyListing;
     const { applicationsByProperty } = state.tenantSelection;
 
-    if (!propertyId) {
-      throw new Error('Property ID is required to send applications to landlord');
-    }
-
-    if (!propertyLandlordId) {
-      throw new Error('Property landlord ID is required to send tenant to landlord');
-    }
-
     // Get tenant applications for the specific property
     const propertyApplications = applicationsByProperty[propertyId] || [];
 
@@ -111,46 +98,16 @@ export const sendAcceptedApplicationsToLandlordAsync = createAsyncThunk(
       app.status === TenantApplicationStatus.ACCEPTED
     );
 
-    if (acceptedApplications.length === 0) {
-      throw new Error('No accepted applications to send to landlord');
-    }
-
-    // Create task for landlord
-    const taskName = `Review ${acceptedApplications.length} Tenant Application(s)`;
-    const taskDescription = `Review ${acceptedApplications.length} accepted tenant application(s) for property at ${streetNumber} ${street}, ${suburb}, ${province} ${postcode}. Applicants: ${acceptedApplications.map((app: TenantApplication) => app.name).join(', ')}`;
-    const dueDate = new Date();
-    dueDate.setDate(dueDate.getDate() + 7);
-
-    // Create task for landlord
-    const taskResult = await apiCreateTaskForLandlord({
-      name: taskName,
-      description: taskDescription,
-      dueDate: dueDate,
-      priority: TaskPriority.MEDIUM,
-      userId: propertyLandlordId,
-      propertyAddress: `${streetNumber} ${street}, ${suburb}, ${province} ${postcode}`,
-      propertyId: propertyId,
-      userType: Role.LANDLORD,
-    });
-
-    // Update all accepted applications to LANDLORD_REVIEW status
-    const acceptedApplicationIds = acceptedApplications.map((app: TenantApplication) => app.id);
-    await updateTenantApplicationStatus(
-      acceptedApplicationIds,
-      TenantApplicationStatus.LANDLORD_REVIEW,
-      2,
-      taskResult // Passes the task ID to link applications to the task
-    );
-
-    console.log(`Successfully sent ${acceptedApplications.length} application(s) to landlord ${propertyLandlordId} for property ${propertyId}`);
-
-    return {
-      success: true,
+    return await sendAcceptedApplicationsToLandlordUseCase(
       propertyId,
-      acceptedApplications: acceptedApplicationIds,
-      applicationCount: acceptedApplications.length,
-      taskId: taskResult
-    };
+      propertyLandlordId,
+      streetNumber,
+      street,
+      suburb,
+      province,
+      postcode,
+      acceptedApplications
+    );
   }
 );
 export const loadTenantApplicationsForPropertyAsync = createAsyncThunk(
